fix: separate intervals frontend
