+++ vite/src/utils/product/productItemUtils.ts
@@ -17,21 +17,6 @@ const itemIsUnlimited = (item: ProductItem) => {
 	return item.included_usage === Infinite;
 };
 
-// The API always populates price_interval for priced items; drop it when it
-// just mirrors the reset interval so "split" means price_interval is set.
-export const normalizeResetInterval = (item: ProductItem): ProductItem => {
-	if (nullish(item.price_interval)) return item;
-
-	const sameInterval = item.price_interval === item.interval;
-	const sameCount =
-		(item.price_interval_count ?? 1) === (item.interval_count ?? 1);
-	if (sameInterval && sameCount) {
-		return { ...item, price_interval: null, price_interval_count: null };
-	}
-
-	return item;
-};
-
 export const getItemType = (item: ProductItem) => {
 	if (isPriceItem(item)) {
 		return ProductItemType.Price;
